Flask SQLite/app/requirements.txt
Flask
mysql-connector
